requirements.txt
python-dotenv
tiktoken
nltk
openai
transformers
torch
pyyaml